Add ImageViewFragments, Model, ViewModel and Repository files
ViewModel is a class designed to hold and manage UI-related data in a lifecycle conscious way. ImageActivityViewModel will support ImageActivity and hold UI related data for it like fetchImage.
ImageRepository has following purpose:
- fetchImages() : Fetch uploaded images from Firebase
- uploadImage() : Upload image to Firebase Storage and Realtime Database
Added ImageDetails model to store data in private fields, and to make the data accessible to the rest of the application through setter and getter methods.
CameraActivity will use CameraX to capture image and save it
ImageListFragment will show uploaded images in a grid
ImageDetailFragment will open full image with zoom in and out functions